docker/Baño/Enchufe/Enchufe/requirements.txt
Flask==1.1.2
paho-mqtt